requirements.txt
matplotlib>=3.7.0
numpy>=1.21.0
pydicom>=2.3.0
Pillow>=9.0.0
pandas>=1.5.0
scipy>=1.9.0
pywebview>=4.0.0 